Move classes to new namespace + fix tests
