+++ src/Resources/config/services.yaml
@@ -5,4 +5,4 @@ services:
 
     Becklyn\RadBundle\:
         resource: ../../*
-        exclude: ../../{Entity,Exception,Pagination/PaginatedList.php,Pagination/Pagination.php,Resources,BecklynRadBundle.php}
+        exclude: ../../{Entity,Exception,Pagination/Data,Resources,BecklynRadBundle.php}
