+++ tests/Pagination/Data/PaginatedListTest.php
@@ -1,16 +1,16 @@
 <?php declare(strict_types=1);
 
-namespace Tests\Becklyn\RadBundle\Pagination;
+namespace Tests\Becklyn\RadBundle\Pagination\Data;
 
-use Becklyn\RadBundle\Pagination\PaginatedList;
-use Becklyn\RadBundle\Pagination\Pagination;
+use Becklyn\RadBundle\Pagination\Data\PaginatedList;
+use Becklyn\RadBundle\Pagination\Data\Pagination;
 use PHPUnit\Framework\TestCase;
 
 class PaginatedListTest extends TestCase
 {
     public function testArray () : void
     {
-        $pagination = new Pagination(1, 10);
+        $pagination = new Pagination(1, 50, 10);
         $list = [1, 2, 3];
 
         $paginatedList = new PaginatedList($list, $pagination);
@@ -22,7 +22,7 @@ class PaginatedListTest extends TestCase
 
     public function testIterator () : void
     {
-        $pagination = new Pagination(1, 10);
+        $pagination = new Pagination(1, 50, 10);
         $list = new \ArrayIterator([1, 2, 3]);
 
         $paginatedList = new PaginatedList($list, $pagination);
